requirements.txt
asgiref==3.5.2
python-dateutil==2.8.2
django==4.1.4
django-crispy-forms==1.14.0
django-registration-redux==2.12
python-dotenv==1.0.0
nepali-datetime==1.0.7
psycopg2==2.9.5
pydantic==1.10.4
sqlparse==0.4.3
typing-extensions==4.4.0
tzdata==2022.7
